Move copyToClipboard and copiableInput to their own module
Fix a cyclic import:
- copyToClipboard needs Tooltip to display the "copied!" message
- Tooltip needs DomUtils to use disableClickPropagation

This was breaking the hash based collectstatic (UmapManifestStaticFilesStorage).
