+++ umap/static/umap/js/modules/domutils.js
@@ -1,7 +1,5 @@
 // Utils that needs the DOM
 import * as Utils from './utils.js'
-import { translate } from './i18n.js'
-import Tooltip from './ui/tooltip.js'
 
 // Mirrors L.DomEvent.disableClickPropagation: prevents clicks on `el` from
 // reaching the Leaflet map below. The `_leaflet_disable_click` flag is still
@@ -11,47 +9,6 @@ export const disableClickPropagation = (el) => {
     el.addEventListener(type, (event) => event.stopPropagation())
   }
   el._leaflet_disable_click = true
-}
-
-export const copyToClipboard = (textToCopy) => {
-  const tooltip = new Tooltip()
-  // https://stackoverflow.com/a/65996386
-  // Navigator clipboard api needs a secure context (https)
-  if (navigator.clipboard && window.isSecureContext) {
-    navigator.clipboard.writeText(textToCopy)
-  } else {
-    // Use the 'out of viewport hidden text area' trick
-    const textArea = document.createElement('textarea')
-    textArea.value = textToCopy
-
-    // Move textarea out of the viewport so it's not visible
-    textArea.style.position = 'absolute'
-    textArea.style.left = '-999999px'
-
-    document.body.prepend(textArea)
-    textArea.select()
-
-    try {
-      document.execCommand('copy')
-    } catch (error) {
-      console.error(error)
-    } finally {
-      textArea.remove()
-    }
-  }
-  tooltip.open({ content: translate('✅ Copied!'), duration: 5000 })
-}
-
-export const copiableInput = (parent, label, value = '') => {
-  const [container, { input, button }] = Utils.loadTemplateWithRefs(`
-    <div class="copiable-input">
-      <label>${label}<input type="text" readOnly value="${value}" data-ref=input /></label>
-      <button type="button" class="icon icon-24 icon-copy" title="${translate('copy')}" data-ref=button></button>
-    </div>
-  `)
-  button.addEventListener('click', () => copyToClipboard(input.value))
-  parent.appendChild(container)
-  return input
 }
 
 // From https://gist.github.com/Accudio/b9cb16e0e3df858cef0d31e38f1fe46f
